Add mypy: https://mypy.readthedocs.io/en/stable/  (#71)
* Add mypy: https://mypy.readthedocs.io/en/stable/ to ensure that we use the correct types and that the code is consistent.

* Flake8

* Various fixes.
Fix DirichletBC typing.
Rename functions in NewtonSolver to python-style functions

* Revert h5py to pypi.

* Flake8

* Fix docs and naming

* Update version number and python support

* Fix build exclude

* Apply suggestions from code review

Revert code that was only tested locally

* Use single array for all coefficients

+++ python/dolfinx_contact/meshing/contact_meshes.py
@@ -78,7 +78,7 @@ def create_circle_plane_mesh(filename: str, quads: bool = False, res=0.1):
     gmsh.finalize()
 
 
-def create_circle_circle_mesh(filename: str, quads: bool = False, res: np.float64 = 0.1):
+def create_circle_circle_mesh(filename: str, quads: bool = False, res: float = 0.1):
     """
     Create two circular meshes, with radii 0.3 and 0.6 with centers (0.5,0.5) and (0.5, -0.5)
     """
